properly handle FF bytes/integer underflow

+++ src/signature.h
@@ -13,8 +13,8 @@ typedef struct {
 } MemoryRegion;
 
 MemoryRegion* get_memory_regions(pid_t pid, int* count);
-bool match_pattern(const uint8_t* data, const int* pattern, size_t pattern_size);
-int* convert_signature(const char* signature, size_t* pattern_size);
+bool match_pattern(const uint8_t* data, const uint16_t* pattern, size_t pattern_size);
+uint16_t* convert_signature(const char* signature, size_t* pattern_size);
 bool validate_process_memory(pid_t pid, uintptr_t address, void* buffer, size_t size);
 int find_signature(lua_State* L);
